add remove dir
