+++ backpack_rs/Cargo.toml
@@ -25,6 +25,7 @@ aws-sdk-s3 = { version = "1.103.0", optional = true, features = ["rustls"] }
 aws-config = { version = "1.8.5", optional = true, features = ["rustls"] }
 aws-credential-types = { version = "1.2.5", optional = true, features = ["hardcoded-credentials"]}
 tokio = { version= "1.47.1" , optional = true, features = ["rt"]}
+tracing-subscriber = "0.3.20"
 
 
 [dev-dependencies]
