+++ backpack_rs/src/fs/file.rs
@@ -18,7 +18,6 @@ pub struct File {
 pub struct OpenOptions {
     inner: StdOpenOptions,
     backend: Arc<RwLock<BoxedBackend>>,
-    must_sync: bool,
 }
 
 impl OpenOptions {
@@ -26,7 +25,6 @@ impl OpenOptions {
         Self {
             inner: StdOpenOptions::new(),
             backend,
-            must_sync: true,
         }
     }
 
@@ -42,6 +40,7 @@ impl OpenOptions {
 
     pub fn append(&mut self, append: bool) -> &mut Self {
         self.inner.append(append);
+
         self
     }
 
@@ -52,7 +51,6 @@ impl OpenOptions {
 
     pub fn create(&mut self, create: bool) -> &mut Self {
         self.inner.create(create);
-        self.must_sync = false;
         self
     }
 
@@ -64,9 +62,9 @@ impl OpenOptions {
     pub fn open<P: AsRef<std::path::Path>>(&mut self, mut path: P) -> std::io::Result<File> {
         let backend = self.backend.read().unwrap();
         let full_path = backend.path().join(path.as_ref());
-        if self.must_sync {
-            // the call initiates downloading the file if it does not exist locally
-            backend.try_exists(path.as_ref())?;
+        if !full_path.exists() {
+            // the call initiates downloading the file from remote storage if needed
+            backend.download(&full_path)?;
         }
 
         let file = self.inner.open(full_path.clone())?;
